Add a `TiffVolume' reader, which reads single-file TIFF image stacks.
In particular this seems to be the approach taken for OME-TIFF
stacks.  Currently works for the sample data I've got, but makes
some bad assumptions.  Two known limitations are it assumes
single-component 8-bit data, and it can't handled compressed
(``Encoded'' in TIFF-lingo) TIFFs.

It also currently ignores all the XML garbage which is stored in
the header for the first file.  All of that information is
accessible via TIFF `tags', so I'm not sure why they shoved the
hard-to-parse-and-interpret XML crud in there at all.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@836 c36c8488-0289-0348-9b64-b301f74bd9a7