Fix cases where we assuming memory is contiguous.
According to 9.2, point (paragraph?) 12, data members within a
class may or may not be contiguous.  This disables the int
operator* in our vectors, to make it more difficult for client
objects pull out the first data member and assume the others
follow immediately.

Array access operators were added to ease the transition.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@891 c36c8488-0289-0348-9b64-b301f74bd9a7