Add the new `Controller'.
This is to provide an easier interface for accessing the master
controller, and most notably logging.

The one MController is now a proper singleton, initialized during
static initialization.  Convenience methods exist in Controller.h
to access that instance; basically, you want to use one of:

  Controller::Instance()
  Controller::ConstInstance()

which return a reference to the MC, with obvious semantics.  For
logging, just use the simplified macros available:

  ERROR("this is an error message");
  WARNING("got back %d instead of 42!", variable);
  MESSAGE("Condition false, looping again...");
  OTHER("Am I ever used?");

These will ensure the proper function name appears in the debug
message.  Assume a printf syntax.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@909 c36c8488-0289-0348-9b64-b301f74bd9a7