Always render `annotations'.
Abstracts out rendering of non-dataset components to pre- and
post- methods.  Call those methods when we start or finish a
subframe.

Previously we'd bail ASAP if all the bricks for the subframe had
been rendered.  This meant that if no bricks were to be rendered
(i.e. they might all be clipped), we'd render nothing -- even if
the user had bounding boxes, etc. enabled.  This was confusing, as
clipping out your entire dataset made your clip plane blink out of
existence (well, visually).  With this change we always render
those planes/boxes/etc., regardless of how many bricks there are
for a given LOD.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@918 c36c8488-0289-0348-9b64-b301f74bd9a7