Fixed a bug in logging while we are dying.
We rely on external knowledge of the deinitialization order; the
comment really says it all.

Maybe this should be redesigned; perhaps the debug output could be
statically initialized as well, and simply grafted into the MC for
API compatibility.  Then we have static initialization order
issues though, which come with their own (albeit solvable)
issues...

I'll investigate after the integration projects are done.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@956 c36c8488-0289-0348-9b64-b301f74bd9a7