Fix the gcc compile error by using a functor.
Seems we hit some strange sort-of partial specialization case,
which the Microsoft compiler seems to support as an extension.
(Guessing, haven't verified that; regardless, gcc didn't support
it ;)

The solution wasn't as simple as changing a template parameter to
a function parameter.  The template instantiation and the function
call location were far removed, and the number of components was
only known at instantiation time.  Thus we couldn't have used a
new function parameter without adding that parameter to
FlatDataToBrickedLOD.

Instead we now use a functor which has a tiny bit of state holding
the number of components.  This allows us to set the value at
instantiation time, as before.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@964 c36c8488-0289-0348-9b64-b301f74bd9a7