Fix the MaxMin issue while retaining the template arguments.
Reverts the earlier fix.  After a quick grep through the spec, it
turns out that default arguments on fqn pointers are illegal.
That gcc only complained when it was given a
two-template-parameter function, and not on the earlier line with
a one-template-parameter function, is apparently a bug.

This removes the default arguments for the function pointers,
forcing a user to explicitly specify them always.  We were never
taking advantage of the default arguments anyway.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@967 c36c8488-0289-0348-9b64-b301f74bd9a7