Search for the amount of memory via /proc first.
A user reported that ImageVis3D wasn't correctly finding the
amount of memory on his system correctly.  A long while back, I
added some code to read /proc for the amount of memory, which
seemed more reliable, but for some reason I only used it as a
fallback.

This changes the code to try /proc first, and only try the
unreliable mechanisms should /proc fail to give us what we want.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1028 c36c8488-0289-0348-9b64-b301f74bd9a7