Use a size_t for the LOD level.
We use the LOD level as an index into arrays at various times.  Since
we were forcing it to be 64bit, this meant win32 spit out a ton of warnings
about our index being larger than a valid array index and was being implicitly
cast.  This makes it a size_t so 32bit platforms get fewer warnings.

Unfortunately this means we can no longer work with datasets which have
more than 4 billion LODs.  Sorry.  If you have such a dataset, I suggest
stealing a supercomputer from the (distant) future to process it.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1060 c36c8488-0289-0348-9b64-b301f74bd9a7