Allow modification of the view parameters.
We still hardcode them by default, but the hardcoding happens in
AbstrRenderer.  When we give the rendering parameters to the
underlying render library, we'll read from variables which are
settable.

ImageVis3D works simply by modifying the preconfigured view
matrices, but other clients need to set this camera information
explicitly.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1079 c36c8488-0289-0348-9b64-b301f74bd9a7