Add checks and workarounds for strange DICOM issues.
This adds some checks to make sure we won't hang when loading some
buggy DICOMs.  With these changes, we get through the GDCM
conformance test without any hangs or crashes.  We cannot load all
of the data, but Tuvok will complain about the poor DICOMs and
then skip over them.

Code was specifically added to crash early in debug mode, to mark
where we need to come back to.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1120 c36c8488-0289-0348-9b64-b301f74bd9a7