External datasets: Avoid a data copy.
Use tr1's shared_ptr to share ownership of external datasets with
the client application.  This means we don't have to copy the
dataset that the application gives us (since we have no idea how
long it'll live), making things considerably faster.

We still do a copy reading the data *out* of the dataset into the
GPU MM.  This will require interface changes here and in the GPU
MM to fix; I'll get to that soon, but want to give some heads up
to other clients first.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1185 c36c8488-0289-0348-9b64-b301f74bd9a7