Fix data size calculation for MaxMinDataBlock.
It wasn't taking into account the number of components in the
data, even though it was writing out max/min info for each
component.

We were also relying on InternalMaxMinElement being tightly
packed.  I removed the reliance on that, but now we have a
dependence on the number of MaxMin elements we write out.  I added
an assert and a big fat comment detailing the danger.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1291 c36c8488-0289-0348-9b64-b301f74bd9a7