Fix the windows signedness warnings in the new minmax code.
At least on Windows, a unary negation operator cannot change the type of a
variable.  As such, our mechanism for computing a bogus initial value for
maximums spewed out a bunch of warnings.

This uses the type tag idiom (i.e. the idea used by iterators) to conditionalize
the initial max value on the signedness of the input type.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1329 c36c8488-0289-0348-9b64-b301f74bd9a7