Fix histogram computation for expressions.
We appeared to have an optimization where we don't re-set parts of
the histogram if they were calculated previously.  However, the
assumption that the old histogram is valid is violated when
computing histograms for expressions; the underlying data has
changed, so the histogram must be thrown out and generated anew.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1619 c36c8488-0289-0348-9b64-b301f74bd9a7