Fix conversion of FP to size_t.
The RH compiler was the only one to catch (somehow?) that we were
the FP range as the number of bins required for a particular
dataset.  Of course, the range of the data has no implication as
to the number of bins needed for FP data.

Add a workaround so that we consider the number of required bins
"huge" when we see FP data.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1647 c36c8488-0289-0348-9b64-b301f74bd9a7