Don't change the source filename between quantize calls.
If we quantize, but then later realize we don't have very many
values and can therefore bin to get a much better data value
distribution... we don't want to overwrite the filename.  That is,
the BinningQuantize should happen on the original data, not the
quantized data.

It is probably feasible to redo the conversion on the quantized
data, but we have to deal with new issues that way.  Most notably,
the size of a datum will have changed during the quantization.
Just redo the work; this isn't the bottleneck anyway.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1665 c36c8488-0289-0348-9b64-b301f74bd9a7