Make data processing functions input from streams.
Not filenames.  This makes them independent of the input files,
which lets the caller take more responsibility in ownership (read:
deletion) of temporary files.

With this we can make temporary files in ConvertRAWDataset and
pass them into the quantization functions, without the
quantization functions needing to know whether the file is
temporary or the original data set.  This in turn simplifies all
the ownership information: when temporary files go out of scope,
they get deleted, so there's no need to "remember" which files
we've created.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1668 c36c8488-0289-0348-9b64-b301f74bd9a7