Don't close the file when we're done quantizing.
Let the caller worry about that.  The callee didn't open the file,
so it shouldn't close it.

This fixes the case of quantizing a file and, while doing so,
realizing we can bin the results to quantize even better.  The
BinningQuantize step assumes the file is open (as it should), so
closing it prematurely makes it error out.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1782 c36c8488-0289-0348-9b64-b301f74bd9a7