*Finally* removed the MemMgr dependence from IO!
When we want to load a data set, the client code makes a call to
IOManager's LoadDataset. That call, in turn, jumps into the GPU
memory manager and asks *it* to load the data set. That memory
manager actually just calls into IOManager to create the DS, it
doesn't do it itself; the only reason the mem manager is involved
is so that it can track how much memory we have loaded.

All this is fine, but it had the one major drawback that IO code
had an explicit dependency on the rendering code, where the GPU
memory manager lived. This really sucks for separating out the IO
code, because it meant you couldn't get by without the renderer,
and therefore OpenGL.

This commit drops the call to GPUMemMan in favor of a single
callback function. There *is* the additional downside that the
client code must set this function pointer before we can load a
data set. But that is outweighed by the non-dependence of the IO
code on GL.

This has been bothering me for literally years. Really didn't take
that much work to fix, either...

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1800 c36c8488-0289-0348-9b64-b301f74bd9a7