Fix largefile using boost types and gcc issue.
I guess the variate_generator stuff is fine in the tr1 namespace in
gcc 4.1, but didn't make it into the c++0x frontend until 4.5.
Just disable that code for now, 4.5 will be standard before we know
it.

If we wanted to, we could always write the code to use
std::tr1::variate_generator on gcc 4.1-4.4, and then use
std::variate_generator on later versions, but this (test) code
isn't important enough right now.


git-svn-id: https://gforge.sci.utah.edu/svn/scio@1835 c36c8488-0289-0348-9b64-b301f74bd9a7