rdb: Test for proper reads/writes.
A while back I was debugging some code that turned out to be
passing invalid filenames around, but it took forever to notice
because RDB happily read nothing and didn't tell me.

No bug right now, but I happened to be reading over this code to
be sure and added these checks, and they're good to have to
prevent a future painful debugging session.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1860 c36c8488-0289-0348-9b64-b301f74bd9a7