geomview: scaling fixes, progress, filename.
Move the data 0.5 back and don't scale it to the unit cube
anymore... rely on the scaling being done outside IV3D for right
now.

Add some (adaptive) progress mechanisms so it doesn't look like
IV3D just hung.

Use the filename for the description, so that the user can
actually distinguish the surfaces when they add a bunch of them to
a volume.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@1870 c36c8488-0289-0348-9b64-b301f74bd9a7