Fix some MSVC compilation problems.
MSVC's std::array initializer only supports the aggregate
initializer syntax (i.e. "{{a,b,c,q}}") if it appears on the RHS
of an assignment(? wtf?).  So avoid it for most cases.

We needed an ifdef in the header because gcc doesn't let us leave
'const' members uninitialized, yet they *can't* be initialized
without the appropriate syntax, so...

git-svn-id: https://gforge.sci.utah.edu/svn/scio@2044 c36c8488-0289-0348-9b64-b301f74bd9a7