Rework how the FileBackedDS/BrickedDS inheritance works.
Remove FileBackedDataset from the chain completely and just make
it a mix-in.  It was always ambiguous whether a FB-DS should
derive from a Bricked-DS, or the other way around; we just decided
one way earlier.
But really the question never made any sense, because it doesn't
matter if a Bricked-DS comes from a file or not.  So the proper
solution is to just consider FileBackedDataset an interface which
some formats can implement and some not.

UVFDataset and DynamicBrickingDS are, in this commit, both
FileBackedDatasets.. but we probably can/should remove that
restriction from the latter.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@2127 c36c8488-0289-0348-9b64-b301f74bd9a7