dynbrick: Save/Load precomputed min/maxes to a cache file.
It creates a dotfile in the current directory, based on the new
bricking params and the original filename.  If such a dotfile
exists it loads the min/max data from that.

This is a huge speedup when loading MM_PRECOMPUTEd rebricked data
sets for the second time.  Inactive for other compute modes.  Has
no effect on rendering, only startup time.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@2173 c36c8488-0289-0348-9b64-b301f74bd9a7