Fix dynbricking rescale issue.
Extend test to actually cover the issue.  Fix the problem, too:
the issue was that we were only overriding the 'RescaleFactors'
methods, but the interface was typically through the more general
'GetScale' method.  So we used our internal scale variables that
we got from derivation (which are otherwise unused, and therefore
nonsense), instead of the variables forwarded from our source DS.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@2196 c36c8488-0289-0348-9b64-b301f74bd9a7