Fix subtle bug converting UVFs that introduced the other day (oops).
I tried to simplify by relying on the shared_ptr to do the
cleanup, but there was an important 'Close' in there, since the
file is re-opened before the function ends (and thus the
shared_ptr cleans up).  Without the close, there's a race between
the OS caching the write and us reading the file's data.

git-svn-id: https://gforge.sci.utah.edu/svn/scio@2211 c36c8488-0289-0348-9b64-b301f74bd9a7